Refactored of headers for TabularDataset (MLDB-1507)
